Sprout - Tenancy + Renaming

+++ sprout/appliances/tasks.py
@@ -24,7 +24,7 @@ import socket
 
 from appliances.models import (
     Provider, Group, Template, Appliance, AppliancePool, DelayedProvisionTask,
-    MismatchVersionMailer, User)
+    MismatchVersionMailer, User, GroupShepherd)
 from sprout import settings, redis
 from sprout.irc_bot import send_message
 from sprout.log import create_logger
@@ -512,7 +512,7 @@ def prepare_template_finish(self, template_id):
     try:
         if template.temporary_name is None:
             tmp_name = "templatize_{}".format(fauxfactory.gen_alphanumeric(8))
-            Template.objects.get(id=template_id).temporary_name = tmp_name
+            Template.objects.get(id=template_id).temporary_name = tmp_name  # metadata, autosave
         else:
             tmp_name = template.temporary_name
         template.provider_api.mark_as_template(
@@ -583,8 +583,7 @@ def apply_lease_times_after_pool_fulfilled(self, appliance_pool_id, time_minutes
     if pool.fulfilled:
         for appliance in pool.appliances:
             apply_lease_times.delay(appliance.id, time_minutes)
-        # TODO: Renaming disabled until orphaning and killing resolved
-        # rename_appliances_for_pool.delay(pool.id)
+        rename_appliances_for_pool.delay(pool.id)
         with transaction.atomic():
             pool.finished = True
             pool.save()
@@ -652,9 +651,7 @@ def replace_clone_to_pool(
     if appliance_pool.not_needed_anymore:
         return
     exclude_template = Template.objects.get(id=exclude_template_id)
-    templates = Template.objects.filter(
-        ready=True, exists=True, usable=True, template_group=appliance_pool.group, version=version,
-        date=date).all()
+    templates = appliance_pool.possible_templates
     templates_excluded = filter(lambda tpl: tpl != exclude_template, templates)
     if templates_excluded:
         template = random.choice(templates_excluded)
@@ -1121,19 +1118,22 @@ def generic_shepherd(self, preconfigured):
     appliances. For each template group, it keeps the last template's appliances spinned up in
     required quantity. If new template comes out of the door, it automatically kills the older
     running template's appliances and spins up new ones. Sorts the groups by the fulfillment."""
-    for grp in sorted(
-            Group.objects.all(), key=lambda g: g.get_fulfillment_percentage(preconfigured)):
+    for gs in sorted(
+            GroupShepherd.objects.all(), key=lambda g: g.get_fulfillment_percentage(preconfigured)):
+        prov_filter = {'provider__user_groups': gs.user_group}
         group_versions = Template.get_versions(
-            template_group=grp, ready=True, usable=True, preconfigured=preconfigured)
+            template_group=gs.template_group, ready=True, usable=True, preconfigured=preconfigured,
+            **prov_filter)
         group_dates = Template.get_dates(
-            template_group=grp, ready=True, usable=True, preconfigured=preconfigured)
+            template_group=gs.template_group, ready=True, usable=True, preconfigured=preconfigured,
+            **prov_filter)
         if group_versions:
             # Downstream - by version (downstream releases)
             version = group_versions[0]
             # Find the latest date (one version can have new build)
             dates = Template.get_dates(
-                template_group=grp, ready=True, usable=True, version=group_versions[0],
-                preconfigured=preconfigured)
+                template_group=gs.template_group, ready=True, usable=True,
+                version=group_versions[0], preconfigured=preconfigured, **prov_filter)
             if not dates:
                 # No template yet?
                 continue
@@ -1151,13 +1151,16 @@ def generic_shepherd(self, preconfigured):
         else:
             continue  # Ignore this group, no templates detected yet
 
+        filter_keep.update(prov_filter)
+        for filt in filters_kill:
+            filt.update(prov_filter)
         # Keeping current appliances
         # Retrieve list of all templates for given group
         # I know joins might be a bit better solution but I'll leave that for later.
         possible_templates = list(
             Template.objects.filter(
-                usable=True, ready=True, template_group=grp, preconfigured=preconfigured,
-                **filter_keep).all())
+                usable=True, ready=True, template_group=gs.template_group,
+                preconfigured=preconfigured, **filter_keep).all())
         # If it can be deployed, it must exist
         possible_templates_for_provision = filter(lambda tpl: tpl.exists, possible_templates)
         appliances = []
@@ -1168,7 +1171,7 @@ def generic_shepherd(self, preconfigured):
         # If we then want to delete some templates, better kill the eldest. status_changed
         # says which one was provisioned when, because nothing else then touches that field.
         appliances.sort(key=lambda appliance: appliance.status_changed)
-        pool_size = grp.template_pool_size if preconfigured else grp.unconfigured_template_pool_size
+        pool_size = gs.template_pool_size if preconfigured else gs.unconfigured_template_pool_size
         if len(appliances) < pool_size and possible_templates_for_provision:
             # There must be some templates in order to run the provisioning
             # Provision ONE appliance at time for each group, that way it is possible to maintain
@@ -1193,16 +1196,19 @@ def generic_shepherd(self, preconfigured):
                 clone_template_to_appliance.delay(appliance.id, None)
         elif len(appliances) > pool_size:
             # Too many appliances, kill the surplus
+            # Only kill those that are visible only for one group. This is necessary so the groups
+            # don't "fight"
             for appliance in appliances[:len(appliances) - pool_size]:
-                self.logger.info("Killing an extra appliance {}/{} in shepherd".format(
-                    appliance.id, appliance.name))
-                Appliance.kill(appliance)
+                if appliance.is_visible_only_in_group(gs.user_group):
+                    self.logger.info("Killing an extra appliance {}/{} in shepherd".format(
+                        appliance.id, appliance.name))
+                    Appliance.kill(appliance)
 
         # Killing old appliances
         for filter_kill in filters_kill:
             for template in Template.objects.filter(
-                    ready=True, usable=True, template_group=grp, preconfigured=preconfigured,
-                    **filter_kill):
+                    ready=True, usable=True, template_group=gs.template_group,
+                    preconfigured=preconfigured, **filter_kill):
                 for a in Appliance.objects.filter(
                         template=template, appliance_pool=None, marked_for_deletion=False):
                     self.logger.info(
@@ -1296,6 +1302,8 @@ def appliance_rename(self, appliance_id, new_name):
         appliance = Appliance.objects.get(id=appliance_id)
     except ObjectDoesNotExist:
         return None
+    if not appliance.provider.allow_renaming:
+        return None
     if appliance.name == new_name:
         return None
     with redis.appliances_ignored_when_renaming(appliance.name, new_name):
@@ -1319,14 +1327,18 @@ def rename_appliances_for_pool(self, pool_id):
             if appliance.provider_api.can_rename
         ]
         for appliance in appliances:
-            new_name = "{}-{}-{}".format(
-                appliance_pool.owner.username,
-                appliance_pool.group.id,
-                appliance.template.date.strftime("%y%m%d")
-            )
-            if appliance.template.version:
-                new_name += "-{}".format(appliance.template.version)
-            new_name += "-{}".format(fauxfactory.gen_alphanumeric(length=4))
+            if not appliance.provider.allow_renaming:
+                continue
+            new_name = '{}_'.format(appliance_pool.owner.username)
+            if appliance.version and not appliance.version.startswith('...'):
+                # CFME
+                new_name += 'cfme_{}_'.format(appliance.version.replace('.', ''))
+            else:
+                # MIQ
+                new_name += 'miq_'
+            new_name += '{}_{}'.format(
+                appliance.template.date.strftime("%y%m%d"),
+                fauxfactory.gen_alphanumeric(length=4))
             appliance_rename.apply_async(
                 countdown=10,  # To prevent clogging with the transaction.atomic
                 args=(appliance.id, new_name))
